requirements.txt
absl-py
gin-config==0.5.0
loguru==0.6.0
matplotlib
nerfacc==0.3.5
numpy==1.23.3
open3d==0.16.0
opencv-python==4.6.0.66
Pillow==9.2.0
rich==12.6.0
tensorboardX
termcolor
torchmetrics==0.10.0
torchtyping==0.1.4
tqdm==4.64.1